refactor: Refactor to separate view and logic
